kernel - Fix rman_make_alignment_flags()
* Fix rman_make_alignment_flags() to work properly up to 2^63.
  Remove assumption about being able to shift a 32-bit integer
  by 32, this is not hardware-portable.

* Note that this function is only used by ancient pccard code

Reported-by: swildner

+++ sys/sys/rman.h
@@ -136,7 +136,7 @@ int	rman_release_resource(struct resource *r);
 struct	resource *rman_reserve_resource(struct rman *rm, u_long start,
 					u_long end, u_long count,
 					u_int flags, device_t dev);
-uint32_t rman_make_alignment_flags(uint32_t size);
+uint32_t rman_make_alignment_flags(size_t size);
 
 #define rman_get_start(r)	((r)->r_start)
 #define rman_get_end(r)		((r)->r_end)
